Create `plasmapy.utils.typing`
